Docusign (examples v0.1.0)

DocuSign REST API

The DocuSign REST API provides you with a powerful, convenient, and simple Web services API for interacting with DocuSign.

Version v2.1

Functions

envelopes__get_envelope(client \\ new(), account_id, envelope_id, query \\ [])

Specs

envelopes__get_envelope(Tesla.Client.t(), binary(), binary(), [opt]) ::
  {:ok, Docusign.Envelope.t()}
  | {:error, Docusign.ErrorDetails.t()}
  | {:error, any()}
when opt: {:advanced_update, binary()} | {:include, binary()}

Gets the status of a single envelope.

Retrieves the overall status for the specified envelope. To get the status of a list of envelopes, use Envelope: listStatusChanges .

Related topics

How to get envelope information

Query parameters

advanced_update: When true, envelope information can be added or modified.

include: Specifies additional information about the envelope to return. Enter a comma-separated list, such as tabs,recipients. Valid values are:

custom_fields: The custom fields associated with the envelope.

documents: The documents associated with the envelope.

attachments: The attachments associated with the envelope.

extensions: Information about the email settings associated with the envelope.

folders: The folder where the envelope exists.

recipients: The recipients associated with the envelope.

powerform: The PowerForms associated with the envelope.

tabs: The tabs associated with the envelope.

payment_tabs: The payment tabs associated with the envelope.

envelopes__post_envelopes(client \\ new(), account_id, envelope_definition, query \\ [])

Specs

envelopes__post_envelopes(
  Tesla.Client.t(),
  binary(),
  Docusign.EnvelopeDefinition.t(),
  [opt]
) ::
  {:ok, Docusign.EnvelopeSummary.t()}
  | {:error, Docusign.ErrorDetails.t()}
  | {:error, any()}
when opt:
       {:cdse_mode, binary()}
       | {:change_routing_order, binary()}
       | {:completed_documents_only, binary()}
       | {:merge_roles_on_draft, binary()}

Creates an envelope.

Creates and sends an envelope or creates a draft envelope. Envelopes are fundamental resources in the DocuSign platform.

With this method you can:

Create and send an envelope with documents, recipients, and tabs.

Create and send an envelope from a template.

Create and send an envelope from a combination of documents and templates.

Create a draft envelope.

When you use this method to create and send an envelope in a single request, the following parameters in the request body (an envelopeDefinition object) are required:

Parameter	Description

status	Set to sent to send the envelope to recipients.<br>Set to created (or don't set at all) to save the envelope as a draft.

emailSubject	The subject of the email used to send the envelope.

documents	The documents to be signed.

recipients	The email addresses of the envelope recipients.

When you create an envelope by using a composite template, you should specify the envelope custom fields in the inline template. Any custom fields that you specify at the root level are ignored.

If the envelope has a workflow definition and the workflowStatus is paused, the envelope will not be sent immediately, even if the envelope's status is sent.

Related topics

There are many ways to use envelopes. You can create and send an envelope with a single API request, or you can use several API requests to create, populate, and send envelopes.

See:	To learn about:

Envelopes	Envelopes, adding documents, tracking, locking, deleting, templates

Documents	Documents, attachments, supplemental documents, authoritative copies, purging

Recipients	Recipients, recipient types, recipient status

Tabs	Tabs, anchoring tabs, custom tabs, payments

The following how-to articles show how to configure this method's envelopeDefinition request body to perform common tasks.

Requesting signatures

How to request a signature by email

How to request a signature by email using a template

How to request a signature by SMS delivery

How to request a signature using a composite template

Requiring authentication

How to require access code authentication for a recipient

How to require SMS authentication for a recipient

How to require phone authentication for a recipient

How to require knowledge-based authentication (KBA) for a recipient

How to require ID verification (IDV) for a recipient

Sending envelopes

How to send an envelope via your app

How to bulk-send envelopes

How to send a request for payment

Setting tab values

How to set envelope tab values

How to set tab values in a template

Applying brands

How to apply a brand to an envelope

How to apply a brand and template to an envelope

Documents, conditional recipients, pausing a workflow

How to attach documents via binary transfer

How to use conditional recipients

How to pause a signature workflow

Query parameters

cdse_mode: Reserved for DocuSign.

change_routing_order: When true, users can define the routing order of recipients while sending documents for signature.

completed_documents_only: Reserved for DocuSign.

merge_roles_on_draft: When set to true, template roles will be merged, and empty recipients will be removed. This parameter applies when you create a draft envelope with multiple templates. (To create a draft envelope, the status field is set to created.)

Note: DocuSign recommends that this parameter should be set to true whenever you create a draft envelope with multiple templates.

new()

Specs

new() :: Tesla.Client.t()

See new/2.

new(middleware, adapter)

Specs

new([Tesla.Client.middleware()], Tesla.Client.adapter()) :: Tesla.Client.t()

Get new API client instance

views__post_envelope_recipient_view(client \\ new(), account_id, envelope_id, recipient_view_request)

Specs

views__post_envelope_recipient_view(
  Tesla.Client.t(),
  binary(),
  binary(),
  Docusign.RecipientViewRequest.t()
) ::
  {:ok, Docusign.EnvelopeViews.t()}
  | {:error, Docusign.ErrorDetails.t()}
  | {:error, any()}

Returns a URL to the recipient view UI.

Returns a URL that enables you to embed the recipient view of the DocuSign UI in your applications. If the recipient is a signer, then the view will provide the signing ceremony.

This method is only used with envelopes in the sent status.

Due to screen space issues, do not use an <iframe> for embedded operations on mobile devices. For iOS devices, DocuSign recommends using a WebView.

Authentication

Your application is responsible for authenticating the identity of the recipient or signer when you use this method. Use the following parameters to record how the recipient was authenticated.

assertionId

authenticationInstant

authenticationMethod

clientUserId

securityDomain

At a minimum, authenticationMethod and clientUserId are required. The information that you provide is included in the envelope's certificate of completion.

Redirects

After the signer completes or ends the signing ceremony, DocuSign redirects the user's browser back to your app via the returnUrl that you supply.

Note: The user may be redirected through various DocuSign subdomains, depending on the region of the account sending the envelope. Please consult the Allowlists for DocuSign eSignature service section in Security for DocuSign eSignature when setting up your allowlists.

The event status parameter

DocuSign appends an event query parameter to the returnUrl with the outcome of the signing ceremony. Your app can use this event parameter to determine the next step for the envelope. Do not fetch the envelope status by using Envelopes: get or a similar method because doing so could break the DocuSign rule against polling.

Note: Because a user can cancel redirection, close their browser after signing, or spoof the landing URL, the returnUrl should not be the single source of truth for envelope status for your integration.

The URL is time-limited

The URL returned in this method's response is valid for one use, and you must use or display it within a couple of minutes after it is generated. After the recipient is redirected to the recipient view, they must interact with the DocuSign system periodically or their session will time out.

Because the URL is time-limited, it should not be stored or sent through email. After you receive it, immediately redirect the user's browser to the URL.

If you want to invite someone to an embedded signing session via email, the email invitation's URL must be to your application. When invoked, your app should request a recipientView URL from DocuSign and then redirect the signer to that URL.

Maintaining State

After the recipient completes the recipient view (or signing ceremony), they are redirected to your application. Your application can recover state information about the transaction by storing information in a cookie, or by including query parameters in the returnUrl field. For example. https://myapp.example.com/docusign_return?myState=12345 When the user is redirected to your app, the event query parameter will be appended. In this example, prevent spoofing by not using a guessable value as the state value.

Revoking the URL

You can revoke this URL by making a DELETE HTTP request to the URL with no request body.

Related topics

How to request a signature by email

How to request a signature through your app

How to request a signature using a composite template

How to send an envelope via your app

How to set envelope tab values

How to set tab values in a template