Plaid (examples v0.1.0)

The Plaid API

The Plaid REST API. Please see https://plaid.com/docs/api for more details.

Version 2020-09-14_1.33.0

Types

access_token()

Specs

access_token() :: binary()

account_selection_cardinality()

Specs

account_selection_cardinality() :: binary()

account_subtype()

Specs

account_subtype() :: binary()

account_type()

Specs

account_type() :: binary()

ach_class()

Specs

ach_class() :: binary()

application_id()

Specs

application_id() :: binary()

asset_report_id()

Specs

asset_report_id() :: binary()

asset_report_pdf_get_response()

Specs

asset_report_pdf_get_response() :: binary()

asset_report_token()

Specs

asset_report_token() :: binary()

bank_initiated_risk_tier()

Specs

bank_initiated_risk_tier() :: integer()

bank_transfer_amount()

Specs

bank_transfer_amount() :: binary()

bank_transfer_direction()

Specs

bank_transfer_direction() :: binary()

bank_transfer_event_type()

Specs

bank_transfer_event_type() :: binary()

bank_transfer_id()

Specs

bank_transfer_id() :: binary()

bank_transfer_network()

Specs

bank_transfer_network() :: binary()

bank_transfer_status()

Specs

bank_transfer_status() :: binary()

bank_transfer_type()

Specs

bank_transfer_type() :: binary()

country_code()

Specs

country_code() :: binary()

customer_initiated_risk_tier()

Specs

customer_initiated_risk_tier() :: integer()

item_id()

Specs

item_id() :: binary()

payment_schedule_interval()

Specs

payment_schedule_interval() :: binary()

products()

Specs

products() :: binary()

request_id()

Specs

request_id() :: binary()

signal_score()

Specs

signal_score() :: integer()

transfer_amount()

Specs

transfer_amount() :: binary()

transfer_authorization_id()

Specs

transfer_authorization_id() :: binary()

transfer_event_type()

Specs

transfer_event_type() :: binary()

transfer_id()

Specs

transfer_id() :: binary()

transfer_network()

Specs

transfer_network() :: binary()

transfer_status()

Specs

transfer_status() :: binary()

transfer_type()

Specs

transfer_type() :: binary()

verification_refresh_status()

Specs

verification_refresh_status() :: binary()

verification_status()

Specs

verification_status() :: binary()

Functions

accounts_balance_get(client \\ new())

Specs

accounts_balance_get(Tesla.Client.t()) ::
  {:ok, Plaid.AccountsGetResponse.t()} | {:error, any()}

Retrieve real-time balance data

The /accounts/balance/get endpoint returns the real-time balance for each of an Item's accounts. While other endpoints may return a balance object, only /accounts/balance/get forces the available and current balance fields to be refreshed rather than cached. This endpoint can be used for existing Items that were added via any of Plaid’s other products. This endpoint can be used as long as Link has been initialized with any other product, balance itself is not a product that can be used to initialize Link.

accounts_get(client \\ new())

Specs

accounts_get(Tesla.Client.t()) ::
  {:ok, Plaid.AccountsGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Retrieve accounts

The /accounts/get endpoint can be used to retrieve information for any linked Item. Note that some information is nullable. Plaid will only return active bank accounts, i.e. accounts that are not closed and are capable of carrying a balance.

This endpoint retrieves cached information, rather than extracting fresh information from the institution. As a result, balances returned may not be up-to-date; for realtime balance information, use /accounts/balance/get instead.

application_get(client \\ new())

Specs

application_get(Tesla.Client.t()) ::
  {:ok, Plaid.ApplicationGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Retrieve information about a Plaid application

Allows financial institutions to retrieve information about Plaid clients for the purpose of building control-tower experiences

asset_report_audit_copy_create(client \\ new())

Specs

asset_report_audit_copy_create(Tesla.Client.t()) ::
  {:ok, Plaid.AssetReportAuditCopyCreateResponse.t()} | {:error, any()}

Create Asset Report Audit Copy

Plaid can provide an Audit Copy of any Asset Report directly to a participating third party on your behalf. For example, Plaid can supply an Audit Copy directly to Fannie Mae on your behalf if you participate in the Day 1 Certainty™ program. An Audit Copy contains the same underlying data as the Asset Report.

To grant access to an Audit Copy, use the /asset_report/audit_copy/create endpoint to create an audit_copy_token and then pass that token to the third party who needs access. Each third party has its own auditor_id, for example fannie_mae. You’ll need to create a separate Audit Copy for each third party to whom you want to grant access to the Report.

asset_report_audit_copy_get(client \\ new())

Specs

asset_report_audit_copy_get(Tesla.Client.t()) ::
  {:ok, Plaid.AssetReportGetResponse.t()} | {:error, any()}

Retrieve an Asset Report Audit Copy

/asset_report/audit_copy/get allows auditors to get a copy of an Asset Report that was previously shared via the /asset_report/audit_copy/create endpoint. The caller of /asset_report/audit_copy/create must provide the audit_copy_token to the auditor. This token can then be used to call /asset_report/audit_copy/create.

asset_report_audit_copy_remove(client \\ new())

Specs

asset_report_audit_copy_remove(Tesla.Client.t()) ::
  {:ok, Plaid.AssetReportAuditCopyRemoveResponse.t()} | {:error, any()}

Remove Asset Report Audit Copy

The /asset_report/audit_copy/remove endpoint allows you to remove an Audit Copy. Removing an Audit Copy invalidates the audit_copy_token associated with it, meaning both you and any third parties holding the token will no longer be able to use it to access Report data. Items associated with the Asset Report, the Asset Report itself and other Audit Copies of it are not affected and will remain accessible after removing the given Audit Copy.

asset_report_create(client \\ new())

Specs

asset_report_create(Tesla.Client.t()) ::
  {:ok, Plaid.AssetReportCreateResponse.t()} | {:error, any()}

Create an Asset Report

The /asset_report/create endpoint initiates the process of creating an Asset Report, which can then be retrieved by passing the asset_report_token return value to the /asset_report/get or /asset_report/pdf/get endpoints.

The Asset Report takes some time to be created and is not available immediately after calling /asset_report/create. When the Asset Report is ready to be retrieved using /asset_report/get or /asset_report/pdf/get, Plaid will fire a PRODUCT_READY webhook. For full details of the webhook schema, see Asset Report webhooks.

The /asset_report/create endpoint creates an Asset Report at a moment in time. Asset Reports are immutable. To get an updated Asset Report, use the /asset_report/refresh endpoint.

asset_report_filter(client \\ new())

Specs

asset_report_filter(Tesla.Client.t()) ::
  {:ok, Plaid.AssetReportFilterResponse.t()} | {:error, any()}

Filter Asset Report

By default, an Asset Report will contain all of the accounts on a given Item. In some cases, you may not want the Asset Report to contain all accounts. For example, you might have the end user choose which accounts are relevant in Link using the Account Select view, which you can enable in the dashboard. Or, you might always exclude certain account types or subtypes, which you can identify by using the /accounts/get endpoint. To narrow an Asset Report to only a subset of accounts, use the /asset_report/filter endpoint.

To exclude certain Accounts from an Asset Report, first use the /asset_report/create endpoint to create the report, then send the asset_report_token along with a list of account_ids to exclude to the /asset_report/filter endpoint, to create a new Asset Report which contains only a subset of the original Asset Report's data.

Because Asset Reports are immutable, calling /asset_report/filter does not alter the original Asset Report in any way; rather, /asset_report/filter creates a new Asset Report with a new token and id. Asset Reports created via /asset_report/filter do not contain new Asset data, and are not billed.

Plaid will fire a PRODUCT_READY webhook once generation of the filtered Asset Report has completed.

asset_report_get(client \\ new())

Specs

asset_report_get(Tesla.Client.t()) ::
  {:ok, Plaid.AssetReportGetResponse.t()} | {:error, any()}

Retrieve an Asset Report

The /asset_report/get endpoint retrieves the Asset Report in JSON format. Before calling /asset_report/get, you must first create the Asset Report using /asset_report/create (or filter an Asset Report using /asset_report/filter) and then wait for the PRODUCT_READY webhook to fire, indicating that the Report is ready to be retrieved.

By default, an Asset Report includes transaction descriptions as returned by the bank, as opposed to parsed and categorized by Plaid. You can also receive cleaned and categorized transactions, as well as additional insights like merchant name or location information. We call this an Asset Report with Insights. An Asset Report with Insights provides transaction category, location, and merchant information in addition to the transaction strings provided in a standard Asset Report.

To retrieve an Asset Report with Insights, call the /asset_report/get endpoint with include_insights set to true.

asset_report_pdf_get(client \\ new())

Specs

asset_report_pdf_get(Tesla.Client.t()) :: {:ok, binary()} | {:error, any()}

Retrieve a PDF Asset Report

The /asset_report/pdf/get endpoint retrieves the Asset Report in PDF format. Before calling /asset_report/pdf/get, you must first create the Asset Report using /asset_report/create (or filter an Asset Report using /asset_report/filter) and then wait for the PRODUCT_READY webhook to fire, indicating that the Report is ready to be retrieved.

The response to /asset_report/pdf/get is the PDF binary data. The request_id is returned in the Plaid-Request-ID header.

View a sample PDF Asset Report.

asset_report_refresh(client \\ new())

Specs

asset_report_refresh(Tesla.Client.t()) ::
  {:ok, Plaid.AssetReportRefreshResponse.t()} | {:error, any()}

Refresh an Asset Report

An Asset Report is an immutable snapshot of a user's assets. In order to "refresh" an Asset Report you created previously, you can use the /asset_report/refresh endpoint to create a new Asset Report based on the old one, but with the most recent data available.

The new Asset Report will contain the same Items as the original Report, as well as the same filters applied by any call to /asset_report/filter. By default, the new Asset Report will also use the same parameters you submitted with your original /asset_report/create request, but the original days_requested value and the values of any parameters in the options object can be overridden with new values. To change these arguments, simply supply new values for them in your request to /asset_report/refresh. Submit an empty string ("") for any previously-populated fields you would like set as empty.

asset_report_remove(client \\ new())

Specs

asset_report_remove(Tesla.Client.t()) ::
  {:ok, Plaid.AssetReportRemoveResponse.t()} | {:error, any()}

Delete an Asset Report

The /item/remove endpoint allows you to invalidate an access_token, meaning you will not be able to create new Asset Reports with it. Removing an Item does not affect any Asset Reports or Audit Copies you have already created, which will remain accessible until you remove them specifically.

The /asset_report/remove endpoint allows you to remove an Asset Report. Removing an Asset Report invalidates its asset_report_token, meaning you will no longer be able to use it to access Report data or create new Audit Copies. Removing an Asset Report does not affect the underlying Items, but does invalidate any audit_copy_tokens associated with the Asset Report.

auth_get(client \\ new())

Specs

auth_get(Tesla.Client.t()) ::
  {:ok, Plaid.AuthGetResponse.t()} | {:error, Plaid.Error.t()} | {:error, any()}

Retrieve auth data

The /auth/get endpoint returns the bank account and bank identification numbers (such as routing numbers, for US accounts) associated with an Item's checking and savings accounts, along with high-level account data and balances when available.

Note: This request may take some time to complete if auth was not specified as an initial product when creating the Item. This is because Plaid must communicate directly with the institution to retrieve the data.

Also note that /auth/get will not return data for any new accounts opened after the Item was created. To obtain data for new accounts, create a new Item.

bank_transfer_balance_get(client \\ new())

Specs

bank_transfer_balance_get(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferBalanceGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Get balance of your Bank Transfer account

Use the /bank_transfer/balance/get endpoint to see the available balance in your bank transfer account. Debit transfers increase this balance once their status is posted. Credit transfers decrease this balance when they are created.

The transactable balance shows the amount in your account that you are able to use for transfers, and is essentially your available balance minus your minimum balance.

Note that this endpoint can only be used with FBO accounts, when using Bank Transfers in the Full Service configuration. It cannot be used on your own account when using Bank Transfers in the BTS Platform configuration.

bank_transfer_cancel(client \\ new())

Specs

bank_transfer_cancel(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferCancelResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Cancel a bank transfer

Use the /bank_transfer/cancel endpoint to cancel a bank transfer. A transfer is eligible for cancelation if the cancellable property returned by /bank_transfer/get is true.

bank_transfer_create(client \\ new())

Specs

bank_transfer_create(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferCreateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Create a bank transfer

Use the /bank_transfer/create endpoint to initiate a new bank transfer.

bank_transfer_event_list(client \\ new())

Specs

bank_transfer_event_list(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferEventListResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

List bank transfer events

Use the /bank_transfer/event/list endpoint to get a list of bank transfer events based on specified filter criteria.

bank_transfer_event_sync(client \\ new())

Specs

bank_transfer_event_sync(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferEventSyncResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Sync bank transfer events

/bank_transfer/event/sync allows you to request up to the next 25 bank transfer events that happened after a specific event_id. Use the /bank_transfer/event/sync endpoint to guarantee you have seen all bank transfer events.

bank_transfer_get(client \\ new())

Specs

bank_transfer_get(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Retrieve a bank transfer

The /bank_transfer/get fetches information about the bank transfer corresponding to the given bank_transfer_id.

bank_transfer_list(client \\ new())

Specs

bank_transfer_list(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferListResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

List bank transfers

Use the /bank_transfer/list endpoint to see a list of all your bank transfers and their statuses. Results are paginated; use the count and offset query parameters to retrieve the desired bank transfers.

bank_transfer_migrate_account(client \\ new())

Specs

bank_transfer_migrate_account(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferMigrateAccountResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Migrate account into Bank Transfers

As an alternative to adding Items via Link, you can also use the /bank_transfer/migrate_account endpoint to migrate known account and routing numbers to Plaid Items. Note that Items created in this way are not compatible with endpoints for other products, such as /accounts/balance/get, and can only be used with Bank Transfer endpoints. If you require access to other endpoints, create the Item through Link instead. Access to /bank_transfer/migrate_account is not enabled by default; to obtain access, contact your Plaid Account Manager.

bank_transfer_sweep_get(client \\ new())

Specs

bank_transfer_sweep_get(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferSweepGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Retrieve a sweep

The /bank_transfer/sweep/get endpoint fetches information about the sweep corresponding to the given sweep_id.

bank_transfer_sweep_list(client \\ new())

Specs

bank_transfer_sweep_list(Tesla.Client.t()) ::
  {:ok, Plaid.BankTransferSweepListResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

List sweeps

The /bank_transfer/sweep/list endpoint fetches information about the sweeps matching the given filters.

categories_get(client \\ new())

Specs

categories_get(Tesla.Client.t()) ::
  {:ok, Plaid.CategoriesGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Get Categories

Send a request to the /categories/get endpoint to get detailed information on categories returned by Plaid. This endpoint does not require authentication.

create_payment_token(client \\ new())

Specs

create_payment_token(Tesla.Client.t()) ::
  {:ok, Plaid.PaymentInitiationPaymentTokenCreateResponse.t()} | {:error, any()}

Create payment token

The /payment_initiation/payment/token/create endpoint has been deprecated. New Plaid customers will be unable to use this endpoint, and existing customers are encouraged to migrate to the newer, link_token-based flow. The recommended flow is to provide the payment_id to /link/token/create, which returns a link_token used to initialize Link.

The /payment_initiation/payment/token/create is used to create a payment_token, which can then be used in Link initialization to enter a payment initiation flow. You can only use a payment_token once. If this attempt fails, the end user aborts the flow, or the token expires, you will need to create a new payment token. Creating a new payment token does not require end user input.

deposit_switch_alt_create(client \\ new())

Specs

deposit_switch_alt_create(Tesla.Client.t()) ::
  {:ok, Plaid.DepositSwitchAltCreateResponse.t()} | {:error, any()}

Create a deposit switch without using Plaid Exchange

This endpoint provides an alternative to /deposit_switch/create for customers who have not yet fully integrated with Plaid Exchange. Like /deposit_switch/create, it creates a deposit switch entity that will be persisted throughout the lifecycle of the switch.

deposit_switch_create(client \\ new())

Specs

deposit_switch_create(Tesla.Client.t()) ::
  {:ok, Plaid.DepositSwitchCreateResponse.t()} | {:error, any()}

Create a deposit switch

This endpoint creates a deposit switch entity that will be persisted throughout the lifecycle of the switch.

deposit_switch_get(client \\ new())

Specs

deposit_switch_get(Tesla.Client.t()) ::
  {:ok, Plaid.DepositSwitchGetResponse.t()} | {:error, any()}

Retrieve a deposit switch

This endpoint returns information related to how the user has configured their payroll allocation and the state of the switch. You can use this information to build logic related to the user's direct deposit allocation preferences.

deposit_switch_token_create(client \\ new())

Specs

deposit_switch_token_create(Tesla.Client.t()) ::
  {:ok, Plaid.DepositSwitchTokenCreateResponse.t()} | {:error, any()}

Create a deposit switch token

In order for the end user to take action, you will need to create a public token representing the deposit switch. This token is used to initialize Link. It can be used one time and expires after 30 minutes.

employers_search(client \\ new())

Specs

employers_search(Tesla.Client.t()) ::
  {:ok, Plaid.EmployersSearchResponse.t()} | {:error, any()}

Search employer database

/employers/search allows you the ability to search Plaid’s database of known employers, for use with Deposit Switch. You can use this endpoint to look up a user's employer in order to confirm that they are supported. Users with non-supported employers can then be routed out of the Deposit Switch flow.

The data in the employer database is currently limited. As the Deposit Switch and Income products progress through their respective beta periods, more employers are being regularly added. Because the employer database is frequently updated, we recommend that you do not cache or store data from this endpoint for more than a day.

identity_get(client \\ new())

Specs

identity_get(Tesla.Client.t()) ::
  {:ok, Plaid.IdentityGetResponse.t()} | {:error, any()}

Retrieve identity data

The /identity/get endpoint allows you to retrieve various account holder information on file with the financial institution, including names, emails, phone numbers, and addresses. Only name data is guaranteed to be returned; other fields will be empty arrays if not provided by the institution.

Note: This request may take some time to complete if identity was not specified as an initial product when creating the Item. This is because Plaid must communicate directly with the institution to retrieve the data.

income_verification_create(client \\ new())

Specs

income_verification_create(Tesla.Client.t()) ::
  {:ok, Plaid.IncomeVerificationCreateResponse.t()} | {:error, any()}

Create an income verification instance

/income/verification/create begins the income verification process by returning an income_verification_id. You can then provide the income_verification_id to /link/token/create under the income_verification parameter in order to create a Link instance that will prompt the user to go through the income verification flow. Plaid will fire an INCOME webhook once the user completes the Payroll Income flow, or when the uploaded documents in the Document Income flow have finished processing.

income_verification_documents_download(client \\ new())

Specs

income_verification_documents_download(Tesla.Client.t()) ::
  {:ok, binary()} | {:error, any()}

Download the original documents used for income verification

/income/verification/documents/download provides the ability to download the source paystub PDF that the end user uploaded via Paystub Import.

The response to /income/verification/documents/download is a ZIP file in binary data. The request_id is returned in the Plaid-Request-ID header.

For Payroll Income, the most recent file available for download with the payroll provider will also be available from this endpoint.

income_verification_paystub_get(client \\ new())

Specs

income_verification_paystub_get(Tesla.Client.t()) ::
  {:ok, Plaid.IncomeVerificationPaystubGetResponse.t()} | {:error, any()}

(Deprecated) Retrieve information from a single paystub used for income verification

income_verification_paystubs_get(client \\ new())

Specs

income_verification_paystubs_get(Tesla.Client.t()) ::
  {:ok, Plaid.IncomeVerificationPaystubsGetResponse.t()} | {:error, any()}

Retrieve information from the paystubs used for income verification

/income/verification/paystubs/get returns the information collected from the paystubs that were used to verify an end user's income. It can be called once the status of the verification has been set to VERIFICATION_STATUS_PROCESSING_COMPLETE, as reported by the INCOME: verification_status webhook. Attempting to call the endpoint before verification has been completed will result in an error.

income_verification_precheck(client \\ new())

Specs

income_verification_precheck(Tesla.Client.t()) ::
  {:ok, Plaid.IncomeVerificationPrecheckResponse.t()} | {:error, any()}

Check a user's eligibility for the income verification product

/income/verification/precheck returns whether a given user is supportable by the income product

income_verification_refresh(client \\ new())

Specs

income_verification_refresh(Tesla.Client.t()) ::
  {:ok, Plaid.IncomeVerificationRefreshResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Refresh an income verification

/income/verification/refresh refreshes a given income verification.

income_verification_summary_get(client \\ new())

Specs

income_verification_summary_get(Tesla.Client.t()) ::
  {:ok, Plaid.IncomeVerificationSummaryGetResponse.t()} | {:error, any()}

Retrieve a summary of information derived from income verification

/income/verification/summary/get returns a verification summary for the income that was verified for an end user. It can be called once the status of the verification has been set to VERIFICATION_STATUS_PROCESSING_COMPLETE, as reported by the INCOME: verification_status webhook. Attempting to call the endpoint before verification has been completed will result in an error.

income_verification_taxforms_get(client \\ new())

Specs

income_verification_taxforms_get(Tesla.Client.t()) ::
  {:ok, Plaid.IncomeVerificationTaxformsGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Retrieve information from the tax documents used for income verification

/income/verification/taxforms/get returns the information collected from taxforms that were used to verify an end user's. It can be called once the status of the verification has been set to VERIFICATION_STATUS_PROCESSING_COMPLETE, as reported by the INCOME: verification_status webhook. Attempting to call the endpoint before verification has been completed will result in an error.

institutions_get(client \\ new())

Specs

institutions_get(Tesla.Client.t()) ::
  {:ok, Plaid.InstitutionsGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Get details of all supported institutions

Returns a JSON response containing details on all financial institutions currently supported by Plaid. Because Plaid supports thousands of institutions, results are paginated.

If there is no overlap between an institution’s enabled products and a client’s enabled products, then the institution will be filtered out from the response. As a result, the number of institutions returned may not match the count specified in the call.

institutions_get_by_id(client \\ new())

Specs

institutions_get_by_id(Tesla.Client.t()) ::
  {:ok, Plaid.InstitutionsGetByIdResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Get details of an institution

Returns a JSON response containing details on a specified financial institution currently supported by Plaid.

institutions_search(client \\ new())

Specs

institutions_search(Tesla.Client.t()) ::
  {:ok, Plaid.InstitutionsSearchResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Search institutions

Returns a JSON response containing details for institutions that match the query parameters, up to a maximum of ten institutions per query.

investments_holdings_get(client \\ new())

Specs

investments_holdings_get(Tesla.Client.t()) ::
  {:ok, Plaid.InvestmentsHoldingsGetResponse.t()} | {:error, any()}

Get Investment holdings

The /investments/holdings/get endpoint allows developers to receive user-authorized stock position data for investment-type accounts.

investments_transactions_get(client \\ new())

Specs

investments_transactions_get(Tesla.Client.t()) ::
  {:ok, Plaid.InvestmentsTransactionsGetResponse.t()} | {:error, any()}

Get investment transactions

The /investments/transactions/get endpoint allows developers to retrieve user-authorized transaction data for investment accounts.

Transactions are returned in reverse-chronological order, and the sequence of transaction ordering is stable and will not shift.

Due to the potentially large number of investment transactions associated with an Item, results are paginated. Manipulate the count and offset parameters in conjunction with the total_investment_transactions response body field to fetch all available investment transactions.

item_access_token_invalidate(client \\ new())

Specs

item_access_token_invalidate(Tesla.Client.t()) ::
  {:ok, Plaid.ItemAccessTokenInvalidateResponse.t()} | {:error, any()}

Invalidate access_token

By default, the access_token associated with an Item does not expire and should be stored in a persistent, secure manner.

You can use the /item/access_token/invalidate endpoint to rotate the access_token associated with an Item. The endpoint returns a new access_token and immediately invalidates the previous access_token.

item_application_list(client \\ new())

Specs

item_application_list(Tesla.Client.t()) ::
  {:ok, Plaid.ItemApplicationListResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

List a user’s connected applications

List a user’s connected applications

item_application_scopes_update(client \\ new())

Specs

item_application_scopes_update(Tesla.Client.t()) ::
  {:ok, Plaid.ItemApplicationScopesUpdateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Update the scopes of access for a particular application

Enable consumers to update product access on selected accounts for an application.

item_create_public_token(client \\ new())

Specs

item_create_public_token(Tesla.Client.t()) ::
  {:ok, Plaid.ItemPublicTokenCreateResponse.t()} | {:error, any()}

Create public token

Note: As of July 2020, the /item/public_token/create endpoint is deprecated. Instead, use /link/token/create with an access_token to create a Link token for use with update mode.

If you need your user to take action to restore or resolve an error associated with an Item, generate a public token with the /item/public_token/create endpoint and then initialize Link with that public_token.

A public_token is one-time use and expires after 30 minutes. You use a public_token to initialize Link in update mode for a particular Item. You can generate a public_token for an Item even if you did not use Link to create the Item originally.

The /item/public_token/create endpoint is not used to create your initial public_token. If you have not already received an access_token for a specific Item, use Link to obtain your public_token instead. See the Quickstart for more information.

item_get(client \\ new())

Specs

item_get(Tesla.Client.t()) ::
  {:ok, Plaid.ItemGetResponse.t()} | {:error, Plaid.Error.t()} | {:error, any()}

Retrieve an Item

Returns information about the status of an Item.

item_import(client \\ new())

Specs

item_import(Tesla.Client.t()) ::
  {:ok, Plaid.ItemImportResponse.t()} | {:error, any()}

Import Item

/item/import creates an Item via your Plaid Exchange Integration and returns an access_token. As part of an /item/import request, you will include a User ID (user_auth.user_id) and Authentication Token (user_auth.auth_token) that enable data aggregation through your Plaid Exchange API endpoints. These authentication principals are to be chosen by you.

Upon creating an Item via /item/import, Plaid will automatically begin an extraction of that Item through the Plaid Exchange infrastructure you have already integrated. This will automatically generate the Plaid native account ID for the account the user will switch their direct deposit to (target_account_id).

item_public_token_exchange(client \\ new())

Specs

item_public_token_exchange(Tesla.Client.t()) ::
  {:ok, Plaid.ItemPublicTokenExchangeResponse.t()} | {:error, any()}

Exchange public token for an access token

Exchange a Link public_token for an API access_token. Link hands off the public_token client-side via the onSuccess callback once a user has successfully created an Item. The public_token is ephemeral and expires after 30 minutes.

The response also includes an item_id that should be stored with the access_token. The item_id is used to identify an Item in a webhook. The item_id can also be retrieved by making an /item/get request.

item_remove(client \\ new())

Specs

item_remove(Tesla.Client.t()) ::
  {:ok, Plaid.ItemRemoveResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Remove an Item

The /item/remove endpoint allows you to remove an Item. Once removed, the access_token associated with the Item is no longer valid and cannot be used to access any data that was associated with the Item.

Note that in the Development environment, issuing an /item/remove request will not decrement your live credential count. To increase your credential account in Development, contact Support.

Also note that for certain OAuth-based institutions, an Item removed via /item/remove may still show as an active connection in the institution's OAuth permission manager.

item_webhook_update(client \\ new())

Specs

item_webhook_update(Tesla.Client.t()) ::
  {:ok, Plaid.ItemWebhookUpdateResponse.t()} | {:error, any()}

Update Webhook URL

The POST /item/webhook/update allows you to update the webhook URL associated with an Item. This request triggers a WEBHOOK_UPDATE_ACKNOWLEDGED webhook to the newly specified webhook URL.

liabilities_get(client \\ new())

Specs

liabilities_get(Tesla.Client.t()) ::
  {:ok, Plaid.LiabilitiesGetResponse.t()} | {:error, any()}

Retrieve Liabilities data

The /liabilities/get endpoint returns various details about an Item with loan or credit accounts. Liabilities data is available primarily for US financial institutions, with some limited coverage of Canadian institutions. Currently supported account types are account type credit with account subtype credit card or paypal, and account type loan with account subtype student or mortgage. To limit accounts listed in Link to types and subtypes supported by Liabilities, you can use the account_filters parameter when creating a Link token.

The types of information returned by Liabilities can include balances and due dates, loan terms, and account details such as original loan amount and guarantor. Data is refreshed approximately once per day; the latest data can be retrieved by calling /liabilities/get.

Note: This request may take some time to complete if liabilities was not specified as an initial product when creating the Item. This is because Plaid must communicate directly with the institution to retrieve the additional data.

link_token_create(client \\ new())

Specs

link_token_create(Tesla.Client.t()) ::
  {:ok, Plaid.LinkTokenCreateResponse.t()} | {:error, any()}

Create Link Token

The /link/token/create endpoint creates a link_token, which is required as a parameter when initializing Link. Once Link has been initialized, it returns a public_token, which can then be exchanged for an access_token via /item/public_token/exchange as part of the main Link flow.

A link_token generated by /link/token/create is also used to initialize other Link flows, such as the update mode flow for tokens with expired credentials, or the Payment Initiation (Europe) flow.

link_token_get(client \\ new())

Specs

link_token_get(Tesla.Client.t()) ::
  {:ok, Plaid.LinkTokenGetResponse.t()} | {:error, any()}

Get Link Token

The /link/token/get endpoint gets information about a previously-created link_token using the /link/token/create endpoint. It can be useful for debugging purposes.

new()

Specs

new() :: Tesla.Client.t()

See new/2.

new(middleware, adapter)

Specs

new([Tesla.Client.middleware()], Tesla.Client.adapter()) :: Tesla.Client.t()

Get new API client instance

payment_initiation_payment_create(client \\ new())

Specs

payment_initiation_payment_create(Tesla.Client.t()) ::
  {:ok, Plaid.PaymentInitiationPaymentCreateResponse.t()} | {:error, any()}

Create a payment

After creating a payment recipient, you can use the /payment_initiation/payment/create endpoint to create a payment to that recipient. Payments can be one-time or standing order (recurring) and can be denominated in either EUR or GBP. If making domestic GBP-denominated payments, your recipient must have been created with BACS numbers. In general, EUR-denominated payments will be sent via SEPA Credit Transfer and GBP-denominated payments will be sent via the Faster Payments network, but the payment network used will be determined by the institution. Payments sent via Faster Payments will typically arrive immediately, while payments sent via SEPA Credit Transfer will typically arrive in one business day.

Standing orders (recurring payments) must be denominated in GBP and can only be sent to recipients in the UK. Once created, standing order payments cannot be modified or canceled via the API. An end user can cancel or modify a standing order directly on their banking application or website, or by contacting the bank. Standing orders will follow the payment rules of the underlying rails (Faster Payments in UK). Payments can be sent Monday to Friday, excluding bank holidays. If the pre-arranged date falls on a weekend or bank holiday, the payment is made on the next working day. It is not possible to guarantee the exact time the payment will reach the recipient’s account, although at least 90% of standing order payments are sent by 6am.

In the Development environment, payments must be below 5 GBP / EUR. For details on any payment limits in Production, contact your Plaid Account Manager.

payment_initiation_payment_get(client \\ new())

Specs

payment_initiation_payment_get(Tesla.Client.t()) ::
  {:ok, any()} | {:error, any()}

Get payment details

The /payment_initiation/payment/get endpoint can be used to check the status of a payment, as well as to receive basic information such as recipient and payment amount. In the case of standing orders, the /payment_initiation/payment/get endpoint will provide information about the status of the overall standing order itself; the API cannot be used to retrieve payment status for individual payments within a standing order.

payment_initiation_payment_list(client \\ new())

Specs

payment_initiation_payment_list(Tesla.Client.t()) ::
  {:ok, Plaid.PaymentInitiationPaymentListResponse.t()} | {:error, any()}

List payments

The /payment_initiation/payment/list endpoint can be used to retrieve all created payments. By default, the 10 most recent payments are returned. You can request more payments and paginate through the results using the optional count and cursor parameters.

payment_initiation_payment_reverse(client \\ new())

Specs

payment_initiation_payment_reverse(Tesla.Client.t()) ::
  {:ok, Plaid.PaymentInitiationPaymentReverseResponse.t()} | {:error, any()}

Reverse an existing payment

Reverse a previously initiated payment.

A payment can only be reversed once and will be refunded to the original sender's account.

payment_initiation_recipient_create(client \\ new())

Specs

payment_initiation_recipient_create(Tesla.Client.t()) ::
  {:ok, Plaid.PaymentInitiationRecipientCreateResponse.t()} | {:error, any()}

Create payment recipient

Create a payment recipient for payment initiation. The recipient must be in Europe, within a country that is a member of the Single Euro Payment Area (SEPA). For a standing order (recurring) payment, the recipient must be in the UK.

The endpoint is idempotent: if a developer has already made a request with the same payment details, Plaid will return the same recipient_id.

payment_initiation_recipient_get(client \\ new())

Specs

payment_initiation_recipient_get(Tesla.Client.t()) ::
  {:ok, any()} | {:error, any()}

Get payment recipient

Get details about a payment recipient you have previously created.

payment_initiation_recipient_list(client \\ new())

Specs

payment_initiation_recipient_list(Tesla.Client.t()) ::
  {:ok, Plaid.PaymentInitiationRecipientListResponse.t()} | {:error, any()}

List payment recipients

The /payment_initiation/recipient/list endpoint list the payment recipients that you have previously created.

processor_apex_processor_token_create(client \\ new())

Specs

processor_apex_processor_token_create(Tesla.Client.t()) ::
  {:ok, Plaid.ProcessorTokenCreateResponse.t()} | {:error, any()}

Create Apex bank account token

Used to create a token suitable for sending to Apex to enable Plaid-Apex integrations.

processor_auth_get(client \\ new())

Specs

processor_auth_get(Tesla.Client.t()) ::
  {:ok, Plaid.ProcessorAuthGetResponse.t()} | {:error, any()}

Retrieve Auth data

The /processor/auth/get endpoint returns the bank account and bank identification number (such as the routing number, for US accounts), for a checking or savings account that's associated with a given processor_token. The endpoint also returns high-level account data and balances when available.

processor_balance_get(client \\ new())

Specs

processor_balance_get(Tesla.Client.t()) ::
  {:ok, Plaid.ProcessorBalanceGetResponse.t()} | {:error, any()}

Retrieve Balance data

The /processor/balance/get endpoint returns the real-time balance for each of an Item's accounts. While other endpoints may return a balance object, only /processor/balance/get forces the available and current balance fields to be refreshed rather than cached.

processor_bank_transfer_create(client \\ new())

Specs

processor_bank_transfer_create(Tesla.Client.t()) ::
  {:ok, Plaid.ProcessorBankTransferCreateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Create a bank transfer as a processor

Use the /processor/bank_transfer/create endpoint to initiate a new bank transfer as a processor

processor_identity_get(client \\ new())

Specs

processor_identity_get(Tesla.Client.t()) ::
  {:ok, Plaid.ProcessorIdentityGetResponse.t()} | {:error, any()}

Retrieve Identity data

The /processor/identity/get endpoint allows you to retrieve various account holder information on file with the financial institution, including names, emails, phone numbers, and addresses.

processor_stripe_bank_account_token_create(client \\ new())

Specs

processor_stripe_bank_account_token_create(Tesla.Client.t()) ::
  {:ok, Plaid.ProcessorStripeBankAccountTokenCreateResponse.t()}
  | {:error, any()}

Create Stripe bank account token

Used to create a token suitable for sending to Stripe to enable Plaid-Stripe integrations. For a detailed guide on integrating Stripe, see Add Stripe to your app.

processor_token_create(client \\ new())

Specs

processor_token_create(Tesla.Client.t()) ::
  {:ok, Plaid.ProcessorTokenCreateResponse.t()} | {:error, any()}

Create processor token

Used to create a token suitable for sending to one of Plaid's partners to enable integrations. Note that Stripe partnerships use bank account tokens instead; see /processor/stripe/bank_account_token/create for creating tokens for use with Stripe integrations.

sandbox_bank_transfer_fire_webhook(client \\ new())

Specs

sandbox_bank_transfer_fire_webhook(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxBankTransferFireWebhookResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Manually fire a Bank Transfer webhook

Use the /sandbox/bank_transfer/fire_webhook endpoint to manually trigger a Bank Transfers webhook in the Sandbox environment.

sandbox_bank_transfer_simulate(client \\ new())

Specs

sandbox_bank_transfer_simulate(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxBankTransferSimulateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Simulate a bank transfer event in Sandbox

Use the /sandbox/bank_transfer/simulate endpoint to simulate a bank transfer event in the Sandbox environment. Note that while an event will be simulated and will appear when using endpoints such as /bank_transfer/event/sync or /bank_transfer/event/list, no transactions will actually take place and funds will not move between accounts, even within the Sandbox.

sandbox_income_fire_webhook(client \\ new())

Specs

sandbox_income_fire_webhook(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxIncomeFireWebhookResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Manually fire an Income webhook

Use the /sandbox/income/fire_webhook endpoint to manually trigger an Income webhook in the Sandbox environment.

sandbox_item_fire_webhook(client \\ new())

Specs

sandbox_item_fire_webhook(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxItemFireWebhookResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Fire a test webhook

The /sandbox/item/fire_webhook endpoint is used to test that code correctly handles webhooks. Calling this endpoint triggers a Transactions DEFAULT_UPDATE webhook to be fired for a given Sandbox Item. If the Item does not support Transactions, a SANDBOX_PRODUCT_NOT_ENABLED error will result.

sandbox_item_reset_login(client \\ new())

Specs

sandbox_item_reset_login(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxItemResetLoginResponse.t()} | {:error, any()}

Force a Sandbox Item into an error state

/sandbox/item/reset_login/ forces an Item into an ITEM_LOGIN_REQUIRED state in order to simulate an Item whose login is no longer valid. This makes it easy to test Link's update mode flow in the Sandbox environment. After calling /sandbox/item/reset_login, You can then use Plaid Link update mode to restore the Item to a good state. An ITEM_LOGIN_REQUIRED webhook will also be fired after a call to this endpoint, if one is associated with the Item.

In the Sandbox, Items will transition to an ITEM_LOGIN_REQUIRED error state automatically after 30 days, even if this endpoint is not called.

sandbox_item_set_verification_status(client \\ new())

Specs

sandbox_item_set_verification_status(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxItemSetVerificationStatusResponse.t()} | {:error, any()}

Set verification status for Sandbox account

The /sandbox/item/set_verification_status endpoint can be used to change the verification status of an Item in in the Sandbox in order to simulate the Automated Micro-deposit flow.

Note that not all Plaid developer accounts are enabled for micro-deposit based verification by default. Your account must be enabled for this feature in order to test it in Sandbox. To enable this features or check your status, contact your account manager or submit a product access Support ticket.

For more information on testing Automated Micro-deposits in Sandbox, see Auth full coverage testing.

sandbox_oauth_select_accounts(client \\ new())

Specs

sandbox_oauth_select_accounts(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxOauthSelectAccountsResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Save the selected accounts when connecting to the Platypus Oauth institution

sandbox_processor_token_create(client \\ new())

Specs

sandbox_processor_token_create(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxProcessorTokenCreateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Create a test Item and processor token

Use the /sandbox/processor_token/create endpoint to create a valid processor_token for an arbitrary institution ID and test credentials. The created processor_token corresponds to a new Sandbox Item. You can then use this processor_token with the /processor/ API endpoints in Sandbox. You can also use /sandbox/processor_token/create with the user_custom test username to generate a test account with custom data.

sandbox_public_token_create(client \\ new())

Specs

sandbox_public_token_create(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxPublicTokenCreateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Create a test Item

Use the /sandbox/public_token/create endpoint to create a valid public_token for an arbitrary institution ID, initial products, and test credentials. The created public_token maps to a new Sandbox Item. You can then call /item/public_token/exchange to exchange the public_token for an access_token and perform all API actions. /sandbox/public_token/create can also be used with the user_custom test username to generate a test account with custom data.

sandbox_transfer_simulate(client \\ new())

Specs

sandbox_transfer_simulate(Tesla.Client.t()) ::
  {:ok, Plaid.SandboxTransferSimulateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Simulate a transfer event in Sandbox

Use the /sandbox/transfer/simulate endpoint to simulate a transfer event in the Sandbox environment. Note that while an event will be simulated and will appear when using endpoints such as /transfer/event/sync or /transfer/event/list, no transactions will actually take place and funds will not move between accounts, even within the Sandbox.

signal_decision_report(client \\ new())

Specs

signal_decision_report(Tesla.Client.t()) ::
  {:ok, Plaid.SignalDecisionReportResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Report whether you initiated an ACH transaction

After calling /signal/evaluate, call /signal/decision/report to report whether the transaction was initiated. This endpoint will return an INVALID_REQUEST error if called a second time with a different value for initiated.

signal_evaluate(client \\ new())

Specs

signal_evaluate(Tesla.Client.t()) ::
  {:ok, Plaid.SignalEvaluateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Evaluate a planned ACH transaction

Use /signal/evaluate to evaluate a planned ACH transaction to get a return risk assessment (such as a risk score and risk tier) and additional risk signals.

In order to obtain a valid score for an ACH transaction, Plaid must have an access token for the account, and the Item must be healthy (receiving product updates) or have recently been in a healthy state. If the transaction does not meet eligibility requirements, an error will be returned corresponding to the underlying cause.

signal_return_report(client \\ new())

Specs

signal_return_report(Tesla.Client.t()) ::
  {:ok, Plaid.SignalReturnReportResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Report a return for an ACH transaction

Call the /signal/return/report endpoint to report a returned transaction that was previously sent to the /signal/evaluate endpoint. Your feedback will be used by the model to incorporate the latest risk trend in your portfolio.

transactions_get(client \\ new())

Specs

transactions_get(Tesla.Client.t()) ::
  {:ok, Plaid.TransactionsGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Get transaction data

The /transactions/get endpoint allows developers to receive user-authorized transaction data for credit, depository, and some loan-type accounts (only those with account subtype student; coverage may be limited). For transaction history from investments accounts, use the Investments endpoint instead. Transaction data is standardized across financial institutions, and in many cases transactions are linked to a clean name, entity type, location, and category. Similarly, account data is standardized and returned with a clean name, number, balance, and other meta information where available.

Transactions are returned in reverse-chronological order, and the sequence of transaction ordering is stable and will not shift. Transactions are not immutable and can also be removed altogether by the institution; a removed transaction will no longer appear in /transactions/get. For more details, see Pending and posted transactions.

Due to the potentially large number of transactions associated with an Item, results are paginated. Manipulate the count and offset parameters in conjunction with the total_transactions response body field to fetch all available transactions.

Data returned by /transactions/get will be the data available for the Item as of the most recent successful check for new transactions. Plaid typically checks for new data multiple times a day, but these checks may occur less frequently, such as once a day, depending on the institution. An Item's status.transactions.last_successful_update field will show the timestamp of the most recent successful update. To force Plaid to check for new transactions, you can use the /transactions/refresh endpoint.

Note that data may not be immediately available to /transactions/get. Plaid will begin to prepare transactions data upon Item link, if Link was initialized with transactions, or upon the first call to /transactions/get, if it wasn't. To be alerted when transaction data is ready to be fetched, listen for the INITIAL_UPDATE and HISTORICAL_UPDATE webhooks. If no transaction history is ready when /transactions/get is called, it will return a PRODUCT_NOT_READY error.

transactions_refresh(client \\ new())

Specs

transactions_refresh(Tesla.Client.t()) ::
  {:ok, Plaid.TransactionsRefreshResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Refresh transaction data

/transactions/refresh is an optional endpoint for users of the Transactions product. It initiates an on-demand extraction to fetch the newest transactions for an Item. This on-demand extraction takes place in addition to the periodic extractions that automatically occur multiple times a day for any Transactions-enabled Item. If changes to transactions are discovered after calling /transactions/refresh, Plaid will fire a webhook: TRANSACTIONS_REMOVED will be fired if any removed transactions are detected, and DEFAULT_UPDATE will be fired if any new transactions are detected. New transactions can be fetched by calling /transactions/get.

Access to /transactions/refresh in Production is specific to certain pricing plans. If you cannot access /transactions/refresh in Production, contact Sales for assistance.

transfer_authorization_create(client \\ new())

Specs

transfer_authorization_create(Tesla.Client.t()) ::
  {:ok, Plaid.TransferAuthorizationCreateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Create a transfer authorization

Use the /transfer/authorization/create endpoint to determine transfer failure risk.

In Plaid's sandbox environment the decisions will be returned as follows:

To approve a transfer, make an authorization request with an amount less than the available balance in the account.

To decline a transfer with the rationale code NSF, the available balance on the account must be less than the authorization amount. See Create Sandbox test data for details on how to customize data in Sandbox.

To decline a transfer with the rationale code RISK, the available balance on the account must be exactly $0. See Create Sandbox test data for details on how to customize data in Sandbox.

To permit a transfer with the rationale code MANUALLY_VERIFIED_ITEM, create an Item in Link through the Same Day Micro-deposits flow.

To permit a transfer with the rationale code LOGIN_REQUIRED, reset the login for an Item.

All username/password combinations other than the ones listed above will result in a decision of permitted and rationale code ERROR.

transfer_cancel(client \\ new())

Specs

transfer_cancel(Tesla.Client.t()) ::
  {:ok, Plaid.TransferCancelResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Cancel a transfer

Use the /transfer/cancel endpoint to cancel a transfer. A transfer is eligible for cancelation if the cancellable property returned by /transfer/get is true.

transfer_create(client \\ new())

Specs

transfer_create(Tesla.Client.t()) ::
  {:ok, Plaid.TransferCreateResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Create a transfer

Use the /transfer/create endpoint to initiate a new transfer.

transfer_event_list(client \\ new())

Specs

transfer_event_list(Tesla.Client.t()) ::
  {:ok, Plaid.TransferEventListResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

List transfer events

Use the /transfer/event/list endpoint to get a list of transfer events based on specified filter criteria.

transfer_event_sync(client \\ new())

Specs

transfer_event_sync(Tesla.Client.t()) ::
  {:ok, Plaid.TransferEventSyncResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Sync transfer events

/transfer/event/sync allows you to request up to the next 25 transfer events that happened after a specific event_id. Use the /transfer/event/sync endpoint to guarantee you have seen all transfer events.

transfer_get(client \\ new())

Specs

transfer_get(Tesla.Client.t()) ::
  {:ok, Plaid.TransferGetResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

Retrieve a transfer

The /transfer/get fetches information about the transfer corresponding to the given transfer_id.

transfer_list(client \\ new())

Specs

transfer_list(Tesla.Client.t()) ::
  {:ok, Plaid.TransferListResponse.t()}
  | {:error, Plaid.Error.t()}
  | {:error, any()}

List transfers

Use the /transfer/list endpoint to see a list of all your transfers and their statuses. Results are paginated; use the count and offset query parameters to retrieve the desired transfers.

webhook_verification_key_get(client \\ new())

Specs

webhook_verification_key_get(Tesla.Client.t()) ::
  {:ok, Plaid.WebhookVerificationKeyGetResponse.t()} | {:error, any()}

Get webhook verification key

Plaid signs all outgoing webhooks and provides JSON Web Tokens (JWTs) so that you can verify the authenticity of any incoming webhooks to your application. A message signature is included in the Plaid-Verification header.

The /webhook_verification_key/get endpoint provides a JSON Web Key (JWK) that can be used to verify a JWT.